pyproject.toml
[project]
name = "job-hunting-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "rapidfuzz>=3.14.3",
    "sqlalchemy>=2.0.48",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.File", "fastapi.Form"]
